+++ src/base/atomicops.h
@@ -283,9 +283,9 @@ inline Atomic64 SeqCst_AtomicExchange(volatile Atomic64* ptr,
 
 inline Atomic64 Relaxed_AtomicIncrement(volatile Atomic64* ptr,
                                         Atomic64 increment) {
-  return increment + std::atomic_fetch_add_explicit(helper::to_std_atomic(ptr),
-                                                    increment,
-                                                    std::memory_order_relaxed);
+  return static_cast<ssize_t>(increment) +
+         std::atomic_fetch_add_explicit(helper::to_std_atomic(ptr), increment,
+                                        std::memory_order_relaxed);
 }
 
 inline Atomic64 Acquire_CompareAndSwap(volatile Atomic64* ptr,
