+++ src/codegen/code-stub-assembler.h
@@ -516,8 +516,12 @@ class V8_EXPORT_PRIVATE CodeStubAssembler
 
   uintptr_t ConstexprUintPtrShl(uintptr_t a, int32_t b) { return a << b; }
   uintptr_t ConstexprUintPtrShr(uintptr_t a, int32_t b) { return a >> b; }
-  intptr_t ConstexprIntPtrAdd(intptr_t a, intptr_t b) { return a + b; }
-  uintptr_t ConstexprUintPtrAdd(uintptr_t a, uintptr_t b) { return a + b; }
+  intptr_t ConstexprIntPtrAdd(intptr_t a, intptr_t b) {
+    return a + static_cast<ssize_t>(b);
+  }
+  uintptr_t ConstexprUintPtrAdd(uintptr_t a, uintptr_t b) {
+    return a + static_cast<ssize_t>(b);
+  }
   intptr_t ConstexprWordNot(intptr_t a) { return ~a; }
   uintptr_t ConstexprWordNot(uintptr_t a) { return ~a; }
 #if defined(__CHERI_PURE_CAPABILITY__)
