+++ src/objects/string.h
@@ -558,7 +558,8 @@ class String : public TorqueGeneratedString<String, Name> {
       const uintptr_t non_one_byte_mask = kUintptrAllBitsSet / 0xFFFF * 0x00FF;
 #endif
       while (chars + sizeof(uintptr_t) <= limit) {
-        if (*reinterpret_cast<const uintptr_t*>(chars) & non_one_byte_mask) {
+        if (*reinterpret_cast<const uintptr_t*>(chars) &
+            static_cast<size_t>(non_one_byte_mask)) {
           break;
         }
         chars += (sizeof(uintptr_t) / sizeof(base::uc16));
